Cast pnoa tensor to float

+++ deep_learning/sentinel_wordlcover_pnoa_vndsm_datamodule.py
@@ -288,7 +288,7 @@ class SentinelWorldCoverPNOAVnDSMDataModule(GeoDataModule):
         if isinstance(batch, dict):
             # move all tensors in your custom data structure to the device
             batch['image'] = batch['image'].to(device)
-            batch['mask'] = batch['mask'].to(device)
+            batch['mask'] = batch['mask'].float().to(device)
         else:
             batch = super().transfer_batch_to_device(batch, device, dataloader_idx)
         return batch
